% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metal.R
\name{as_mtl_buffer}
\alias{as_mtl_buffer}
\alias{as_mtl_buffer.mtl_buffer}
\alias{as_mtl_buffer.mtl_floats}
\title{Create buffers}
\usage{
as_mtl_buffer(x, ...)

\method{as_mtl_buffer}{mtl_buffer}(x, ...)

\method{as_mtl_buffer}{mtl_floats}(x, ..., device = mtl_default_device())
}
\arguments{
\item{x}{An object to convert to a buffer}

\item{...}{Passed to S3 methods}

\item{device}{A \link[=mtl_default_device]{mtl_device}}
}
\value{
An object of class 'mtl_buffer'
}
\description{
Create buffers
}
\examples{
as_mtl_buffer(as_mtl_floats(1:10))

}
