% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{default_device}
\alias{default_device}
\title{Get the default metal device}
\usage{
default_device()
}
\value{
An external pointer of class metalme_device
}
\description{
Get the default metal device
}
\examples{
default_device()

}
