% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalme-package.R
\docType{package}
\name{metal-package}
\alias{metal}
\alias{metal-package}
\title{metal: See If We Can Use Apple Metal In R}
\description{
Sees if we can use Apple Metal from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/metal}
  \item Report bugs at \url{https://github.com/paleolimbot/metal/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
