% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalme-package.R
\docType{package}
\name{metalme-package}
\alias{metalme}
\alias{metalme-package}
\title{metalme: See If We Can Use Apple Metal In R}
\description{
Sees if we can use Apple Metal from R
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
