% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metal.R
\name{mtl_buffer}
\alias{mtl_buffer}
\title{Create buffers}
\usage{
mtl_buffer(size, device = mtl_default_device())
}
\arguments{
\item{size}{The size of the buffer in bytes}

\item{device}{A \link[=mtl_default_device]{mtl_device}}

\item{...}{Passed to S3 methods}
}
\value{
An object of class 'mtl_buffer'
}
\description{
Create buffers
}
\examples{
mtl_buffer(1024)

}
