% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metal.R
\name{mtl_compute_pipeline}
\alias{mtl_compute_pipeline}
\alias{mtl_compute_pipeline_execute}
\title{Compile and execute compute functions}
\usage{
mtl_compute_pipeline(func)

mtl_compute_pipeline_execute(
  pipeline,
  length,
  ...,
  device = mtl_default_device()
)
}
\arguments{
\item{func}{An mtl_function}

\item{pipeline}{A pipeline created with \code{\link[=mtl_compute_pipeline]{mtl_compute_pipeline()}}}

\item{length}{The array length to execute across (used to create the grid
of threads)}

\item{...}{Arguments (currently all \code{\link[=mtl_buffer]{mtl_buffer()}}s) or objects that
will be coerced to them.}

\item{device}{A \link[=mtl_default_device]{mtl_device}}
}
\value{
\itemize{
\item \code{mtl_compute_pipeline()} returns an mtl_compute_pipline object representing
a compiled version of the function for the device's GPU.
\item \code{mtl_compute_pipeline_execute()} returns nothing (usually the function
populates an output buffer that is one of the arguments).
}
}
\description{
Compile and execute compute functions
}
