% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metal.R
\name{mtl_default_device}
\alias{mtl_default_device}
\title{Get the default metal device}
\usage{
mtl_default_device()
}
\value{
An external pointer of class metal_device
}
\description{
Get the default metal device
}
\examples{
mtl_default_device()

}
