% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floats.R
\name{mtl_floats}
\alias{mtl_floats}
\alias{as_mtl_floats}
\title{Create float vectors}
\usage{
mtl_floats(size = 0, fill = NA_real_)

as_mtl_floats(x, ...)
}
\arguments{
\item{size}{The size of the buffer to allocate}

\item{fill}{An optional value with which each element should be initialized}

\item{x}{An object to convert to an \code{\link[=mtl_floats]{mtl_floats()}}.}

\item{...}{Passed to S3 methods}
}
\value{
An vctr of class mtl_floats
}
\description{
Create float vectors
}
\examples{
mtl_floats(5, NaN)
as_mtl_floats(1:5)

}
