% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metal.R
\name{mtl_make_library}
\alias{mtl_make_library}
\title{Create a metal function lirary}
\usage{
mtl_make_library(code, device = mtl_default_device())
}
\arguments{
\item{code}{Code in the metal shading language}

\item{device}{A \link[=mtl_default_device]{mtl_device}}
}
\value{
An external pointer of class mtl_library
}
\description{
Create a metal function lirary
}
\examples{
mtl_make_library("
  kernel void add_arrays(device const float* inA,
                         device const float* inB,
                         device float* result,
                         uint index [[thread_position_in_grid]]) {
    result[index] = inA[index] + inB[index];
  }
")

}
